+++ docs/HISTORY.rst
@@ -12,6 +12,7 @@ Change Log
 - Scheduler can now run at much higher speeds if your hardware can take it
 - ``listen_log()`` now sends AppDaemon system messages and has the option to set a log level.
 - Bumped aiohttp to v3.4.4
+- Added callback locking decorator
 
 **Fixes**